requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
ecdsa==0.18.0
fastapi==0.92.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyMySQL==1.0.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
